// src/types/api/player.ts
export type Player = {
  search_player_all: {
    copyRight: string;
    queryResults: {
      created: string;
      totalSize: string;
      row: {
        position: string;
        birth_country: string;
        weight: string;
        birth_state: string;
        name_display_first_last: string;
        college: string;
        height_inches: string;
        name_display_roster: string;
        sport_code: string;
        bats: string;
        name_first: string;
        team_code: string;
        birth_city: string;
        height_feet: string;
        pro_debut_date: string;
        team_full: string;
        team_abbrev: string;
        birth_date: string;
        throws: string;
        league: string;
        name_display_last_first: string;
        position_id: string;
        high_school: string;
        name_use: string;
        player_id: string;
        name_last: string;
        team_id: string;
        service_years: string;
        active_sw: string
        }
      }
    }
  }